+++ src/main/kotlin/com/yapp/yapp/record/domain/RecordService.kt
@@ -2,7 +2,7 @@ package com.yapp.yapp.record.domain
 
 import com.yapp.yapp.record.api.response.RecordListResponse
 import com.yapp.yapp.record.api.response.RecordResponse
-import com.yapp.yapp.running.domain.record.RunningRecordManager
+import com.yapp.yapp.record.domain.record.RunningRecordManager
 import org.springframework.data.domain.Pageable
 import org.springframework.stereotype.Service
 import java.time.OffsetDateTime
