+++ src/main/kotlin/com/yapp/yapp/record/domain/point/RunningPointRepository.kt
@@ -1,6 +1,6 @@
-package com.yapp.yapp.running.domain.point
+package com.yapp.yapp.record.domain.point
 
-import com.yapp.yapp.running.domain.record.RunningRecord
+import com.yapp.yapp.record.domain.record.RunningRecord
 import org.springframework.data.repository.CrudRepository
 
 interface RunningPointRepository : CrudRepository<RunningPoint, Long> {
