+++ src/main/kotlin/com/yapp/yapp/record/domain/record/RunningRecordRepository.kt
@@ -1,4 +1,4 @@
-package com.yapp.yapp.running.domain.record
+package com.yapp.yapp.record.domain.record
 
 import org.springframework.data.domain.Pageable
 import org.springframework.data.repository.CrudRepository
