+++ src/main/kotlin/com/yapp/yapp/running/api/response/RunningDoneResponse.kt
@@ -1,6 +1,6 @@
 package com.yapp.yapp.running.api.response
 
-import com.yapp.yapp.running.domain.record.RunningRecord
+import com.yapp.yapp.record.domain.record.RunningRecord
 import java.time.Duration
 import java.time.OffsetDateTime
