+++ src/test/kotlin/com/yapp/yapp/running/RunningMetricsCalculatorTest.kt
@@ -1,9 +1,9 @@
 package com.yapp.yapp.running
 
 import com.yapp.yapp.common.TimeProvider
+import com.yapp.yapp.record.domain.point.RunningPoint
+import com.yapp.yapp.record.domain.record.RunningRecord
 import com.yapp.yapp.running.domain.RunningMetricsCalculator
-import com.yapp.yapp.running.domain.point.RunningPoint
-import com.yapp.yapp.running.domain.record.RunningRecord
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
